market order

+++ broker/broker_facade/domain/model/facade_order.go
@@ -5,6 +5,6 @@ type FacadeOrder struct {
 	Quantity     int     `json:"quantity"`
 	OrderType    string  `json:"orderType"`
 	OrderSubtype string  `json:"orderSubtype"`
-	OrderPrice   float32 `json:"orderPrice"`
+	OrderPrice   float64 `json:"orderPrice"`
 	Id           string  `json:"id"`
 }
