+++ broker/broker_facade/domain/model/internal_order.go
@@ -5,7 +5,7 @@ type InternalOrder struct {
 	Quantity     int     `json:"quantity"`
 	OrderType    string  `json:"orderType"`
 	OrderSubtype string  `json:"orderSubtype"`
-	OrderPrice   float32 `json:"orderPrice"`
+	OrderPrice   float64 `json:"orderPrice"`
 	ClientId     string  `json:"clientId"`
 	Id           string  `json:"id"`
 }
