+++ broker/order_executor/domain/model/stock_order.go
@@ -5,7 +5,7 @@ type StockOrder struct {
 	Quantity     int     `json:"quantity"`
 	OrderType    string  `json:"orderType"`
 	OrderSubtype string  `json:"orderSubtype"`
-	OrderPrice   float32 `json:"orderPrice"`
+	OrderPrice   float64 `json:"orderPrice"`
 	ClientId     string  `json:"clientId"`
 	BrokerId     string  `json:"brokerId"`
 	Id           string  `json:"id"`
